backend/requirements.txt
fastapi==0.52.0
pymongo==4.3.3
uvicorn==0.22.0
PyJWT==2.7.0
pydantic
mongoengine